// src/model/DashBoardSections.ts
export type DashboardSection =
    | "home"
    | "business"
    | "appearance"
    | "content"
    | "tools"
    | "share"
    | "account";

export type DashboardSubSection =
    "home"
    | "business-information"
    | "socials"
    | "theme"
    | "branding"
    | "faqs"
    | "services"
    | "media"
    | "gallery"
    | "documents"
    | "opening-hours"
    | "enquiries"
    | "updates"
    | "banner"
    | "share"
    | "settings"
    | "account"
    | "help";